% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_soilgrids.R
\name{aggregate_soilgrids}
\alias{aggregate_soilgrids}
\title{Aggregate SoilGrids rasters}
\usage{
aggregate_soilgrids(r)
}
\arguments{
\item{r}{SpatRaster, single/multiple layer with soil characteristics.
Outcome of the \code{\link[=get_soilgrids]{get_soilgrids()}} function.}
}
\value{
SpatRaster
}
\description{
It allows to transform and aggregate SoilGrids rasters. I.e. from
SoilGrids one would download several layers (0-5, 5-15, 15-30, etc.).
This function will take an average among them and convert to conventional
units (see \href{https://www.isric.org/explore/soilgrids/faq-soilgrids#What_do_the_filename_codes_mean}{SoilGrids FAQ})
}
\examples{
library(purrr)

f <- system.file("extdata/extent.shp", package="rusleR")
v <- vect(f)

sand <- get_soilgrids(v, layer = "sand")
sand

sand_agg <- aggregate_soilgrids(sand)
sand_agg

}
