% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glored.R
\name{get_glored}
\alias{get_glored}
\title{Download Global Rainfall Erosivity}
\usage{
get_glored(aoi, warp = TRUE)
}
\arguments{
\item{aoi}{SpatVector. A polygon layer with area of interest.}

\item{warp}{logical. If TRUE, reproject the GLORED raster to \code{aoi} projection}
}
\value{
SpatRaster
}
\description{
This function download Global Rainfall Erosivity Database (GLORED) created by \emph{Panagos et al.} (2017) and crop it to the shapefile boundaries.
}
\examples{
library(terra)

f <- system.file("extdata/extent.shp", package="rusleR")
v <- vect(f)

r_factor <- get_glored(v)

plot(v)

}
\references{
Panagos, Panos, Pasquale Borrelli, Katrin Meusburger, Bofu Yu, Andreas Klik, Kyoung Jae Lim, Jae E. Yang, et al. “Global Rainfall Erosivity Assessment Based on High-Temporal Resolution Rainfall Records.” Scientific Reports 7, no. 1 (June 23, 2017): 4175. https://doi.org/10.1038/s41598-017-04282-8.
}
