% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_factor.R
\name{k_factor}
\alias{k_factor}
\title{Calculate K-factor}
\usage{
k_factor(sand, silt, clay, soc, method = c("williams1983"))
}
\arguments{
\item{sand}{SpatRaster. Single layer raster with sand proportion (\%).
Outcome of the \code{\link[=aggregate_soilgrids]{aggregate_soilgrids()}}}

\item{silt}{SpatRaster. Single layer raster with silt proportion (\%).
Outcome of the \code{\link[=aggregate_soilgrids]{aggregate_soilgrids()}}}

\item{clay}{SpatRaster. Single layer raster with clay proportion (\%).
Outcome of the \code{\link[=aggregate_soilgrids]{aggregate_soilgrids()}}}

\item{soc}{SpatRaster. Single layer raster with weight percentage of organic carbon (\%).
Outcome of the \code{\link[=aggregate_soilgrids]{aggregate_soilgrids()}}}

\item{method}{string. One or more of these options: \code{'williams1983'} (see Details)}
}
\value{
SpatRaster
}
\description{
Calculate K-factor
}
\details{
Currently, this function supports the following equations:
\itemize{
\item \code{'williams1983'} -- An equation from Williams and Renard (1983) as cited in Chen et al. (2011)
}
}
\examples{
library(purrr)

f <- system.file("extdata/extent.shp", package="rusleR")
v <- vect(f)

sand <- get_soilgrids(v, layer = "sand")
silt <- get_soilgrids(v, layer = "silt")
clay <- get_soilgrids(v, layer = "clay")
soc <- get_soilgrids(v, layer = "soc")

sand_mean <- aggregate_soilgrids(sand)
silt_mean <- aggregate_soilgrids(silt)
clay_mean <- aggregate_soilgrids(clay)
soc_mean <- aggregate_soilgrids(soc)

k <- k_factor(sand_mean, silt_mean, clay_mean, soc_mean)
k

}
\references{
Williams, J. R., K. G. Renard, and P. T. Dyke. “EPIC: A New Method for Assessing Erosion’s Effect on Soil Productivity.” Journal of Soil and Water Conservation 38, no. 5 (September 1, 1983): 381–83.

Chen, Liangang, Xin Qian, and Yong Shi. “Critical Area Identification of Potential Soil Loss in a Typical Watershed of the Three Gorges Reservoir Region.” Water Resources Management 25, no. 13 (June 25, 2011): 3445. https://doi.org/10.1007/s11269-011-9864-4.
}
