% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_2d.R
\name{ls_2d}
\alias{ls_2d}
\title{\eqn{LS_{2D}}}
\usage{
ls_2d(.dem, .weights, .threshold = 120, saga_obj = saga)
}
\arguments{
\item{.dem}{SpatRaster. A single layer with elevation values. Values should have the same unit as the map units, or in meters when the crs is longitude/latitude}

\item{.weights}{SpatRaster.}

\item{.threshold}{Numeric. A value specifying a flow threshold in meters. Default is 120 m}

\item{saga_obj}{SAGA-GIS geoprocessor object from \code{\link[Rsagacmd]{saga_gis}} with \code{raster_backend = "terra"}}
}
\value{
SpatRaster
}
\description{
This function calculates an LS-factor from WATEM/SEDEM
model as described in \emph{Batista et al.} (2022).
}
\references{
Batista, P. V. G., Fiener, P., Scheper, S., and Alewell, C.: Data and code for: A conceptual model-based sediment connectivity assessment for patchy agricultural catchments, Zenodo, https://doi.org/10.5281/zenodo.6560226, 2022.
}
