% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_alpine.R
\name{ls_alpine}
\alias{ls_alpine}
\title{\eqn{LS_{Alpine}}}
\usage{
ls_alpine(dem, threshold = 120, saga_obj = saga)
}
\arguments{
\item{dem}{SpatRaster. A single layer with elevation values. Values should have the same unit as the map units, or in meters when the crs is longitude/latitude}

\item{threshold}{Numeric. Value specifying a flow threshold in meters. Default is 120 m}

\item{saga_obj}{SAGA-GIS geoprocessor object. Run \code{Rsagacmd::saga_gis(raster_backend = "terra")}}
}
\value{
SpatRaster
}
\description{
This function calculates an LS-factor from Universal Soil
Loss Equation (USLE), as described in \emph{Schmidt et al.} (2019).
}
\examples{
\dontrun{
library(Rsagacmd)

# initiate a saga object
saga <- saga_gis(raster_backend = "terra")

# load DEM
f <- system.file("extdata/dem.tif", package="rusleR")
DEM <- rast(f)

# calculate LS-alpine
ls <- ls_alpine(dem = DEM)

plot(ls)
}


}
\references{
Schmidt, Simon, Simon Tresch, and Katrin Meusburger. “Modification of the RUSLE Slope Length and Steepness Factor (LS-Factor) Based on Rainfall Experiments at Steep Alpine Grasslands.” MethodsX 6 (2019): 219–29. https://doi.org/10.1016/j.mex.2019.01.004.
}
