% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusleR-package.R
\docType{package}
\name{rusleR-package}
\alias{rusleR}
\alias{rusleR-package}
\title{rusleR: Universal Soil Loss Equation Estimation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package offers an R implementation of Universal Soil Loss Equation (USLE).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://atsyplenkov.github.io/rusleR/}
}

}
\author{
\strong{Maintainer}: Anatoly Tsyplenkov \email{atsyplenkov@gmail.com} (\href{https://orcid.org/0000-0003-4144-8402}{ORCID})

Authors:
\itemize{
  \item Simon Scheper (\href{https://orcid.org/0000-0001-8097-673X}{ORCID})
}

}
\keyword{internal}
