% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watem_erosion.R
\name{watem_erosion}
\alias{watem_erosion}
\title{Calculates water erosion}
\usage{
watem_erosion(.r_factor, .k_factor, .c_factor, .ls_factor)
}
\arguments{
\item{.r_factor}{SpatRaster or Numeric. Rainfall erosivity from RUSLE but
divided by \code{10000} to get \eqn{MJ·mm·m^{−2}·h^{−1}·yr^{−1}}.}

\item{.k_factor}{SpatRaster or Numeric. Soil erodibility from RUSLE but
multiplied by \code{1000} to get \eqn{kg·h·MJ^{−1}·mm^{−1}}.}

\item{.c_factor}{SpatRaster. It is a cover management factor (dimensionless)}

\item{.ls_factor}{SpatRaster. \eqn{LS} factor. An output from \code{\link[rusleR]{ls_2d}}}
}
\value{
SpatRaster.
}
\description{
Calculates water erosion
}
