% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watem_flux.R
\name{watem_flux}
\alias{watem_flux}
\title{WATEM/SEDEM}
\usage{
watem_flux(.dem, .erosion, .transport_capacity, saga_obj = saga)
}
\arguments{
\item{.dem}{SpatRaster.}

\item{.erosion}{SpatRaster.}

\item{.transport_capacity}{SpatRaster.}

\item{saga_obj}{SAGA-GIS geoprocessor object from \code{\link[Rsagacmd]{saga_gis}} with \code{raster_backend = "terra"}}
}
\value{
List.
}
\description{
WATEM/SEDEM
}
