% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watem_ktc.R
\name{watem_ktc}
\alias{watem_ktc}
\title{Transport capacity coefficient}
\usage{
watem_ktc(
  .c_factor,
  .ktc_low = 75,
  .ktc_high = 250,
  .c_max = 0.1,
  .seed = 1234
)
}
\arguments{
\item{.c_factor}{SpatRaster. It is a cover management factor (dimensionless)}

\item{.ktc_low}{Numeric. The lower limit of possible \eqn{K_{TC}}}

\item{.ktc_high}{Numeric. The upper limit of \eqn{K_{TC}}}

\item{.c_max}{Numeric. A threshold C-factor value needs to be
given to indicate for which areas the high \eqn{K_{TC}} needs to
 be used. Usually, roads are given a very high \eqn{K_{TC}} such
 that no sediment deposition is modelled on road surfaces. The default is \code{0.1}.}

\item{.seed}{Numeric.}
}
\value{
SpatRaster
}
\description{
Calculates \eqn{K_{TC}} from WATEM/SEDEM (Van Oost et al., 2000). It is
a land-use-dependent transport capacity coefficient (m). The transport
capacity is the maximum sediment mass that can be transported by the overland
 flow. If the sediment production is higher than this transport capacity,
 sediment will be deposited. Thus, the higher the transport capacity coefficient,
 the more sediment can be transported downslope (Van Oost et al., 2000).
}
\details{
The \eqn{K_{TC}} usually requires calibration (\emph{cf.} Batista et al., 2022).
Here we used an approach suggested in Batista et al. (2022). \eqn{K_{TC}}
value is sampled from an uniform distribution using the \code{\link[stats]{runif}}
function.
}
\references{
Batista PVG, Fiener P, Scheper S, Alewell C. 2022. A conceptual-model-based sediment connectivity assessment for patchy agricultural catchments. Hydrology and Earth System Sciences 26 : 3753–3770. DOI: 10.5194/hess-26-3753-2022

Van Oost K, Govers G, Desmet P. 2000. Evaluating the effects of changes in landscape structure on soil erosion by water and tillage. Landscape Ecology 15 : 577–589. DOI: 10.1023/A:1008198215674
}
