% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watem_ptef.R
\name{watem_ptef}
\alias{watem_ptef}
\title{Parcel trapping efficiency}
\usage{
watem_ptef(.weights, .m, .include.lowest = TRUE)
}
\arguments{
\item{.weights}{SpatRaster. Landuse/Landcover map}

\item{.m}{Vector. A reclassification vector used in
\code{\link[terra]{classify}}}

\item{.include.lowest}{Logical.}
}
\value{
SpatRaster
}
\description{
Parcel trapping efficiency \eqn{P_{TEF}} parameter from
WATEM/SEDEM (Van Oost et al., 2000). It corresponds to the proportion of
the flow accumulation that is routed downstream (Batista et al., 2022).
}
\references{
Batista PVG, Fiener P, Scheper S, Alewell C. 2022. A conceptual-model-based sediment connectivity assessment for patchy agricultural catchments. Hydrology and Earth System Sciences 26 : 3753–3770. DOI: 10.5194/hess-26-3753-2022

Van Oost K, Govers G, Desmet P. 2000. Evaluating the effects of changes in landscape structure on soil erosion by water and tillage. Landscape Ecology 15 : 577–589. DOI: 10.1023/A:1008198215674
}
