% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watem_tc.R
\name{watem_tc}
\alias{watem_tc}
\title{Transport capacity}
\usage{
watem_tc(.dem, .ls, .r_factor, .k_factor, .ktc, saga_obj = saga)
}
\arguments{
\item{.dem}{SpatRaster. A single layer with elevation values.
Values should have the same unit as the map units, or in meters when
the crs is longitude/latitude.}

\item{.ls}{SpatRaster. An output from \code{\link[rusleR]{ls_2d}}}

\item{.r_factor}{SpatRaster or Numeric. Rainfall erosivity from RUSLE but
divided by \code{10000} to get \eqn{MJ·mm·m^{−2}·h^{−1}·yr^{−1}}.}

\item{.k_factor}{SpatRaster or Numeric. Soil erodibility from RUSLE but
multiplied by \code{1000} to get \eqn{kg·h·MJ^{−1}·mm^{−1}}.}

\item{.ktc}{SpatRaster. Tranport capacity coefficient \eqn{K_{TC}}.
An output from \code{\link[rusleR]{watem_ktc}}}

\item{saga_obj}{SAGA-GIS geoprocessor object from \code{\link[Rsagacmd]{saga_gis}} with \code{raster_backend = "terra"}}
}
\value{
SpatRaster
}
\description{
Transport capacity (\eqn{kg·m^{−1}·yr^{−1}}) per unit width of the grid
cells is assumed to be proportional to the potential to rill erosion,
which is described by a power function of slope length and gradient
(Van Rompaey et al., 2001).
}
\references{
Van Rompaey, A., Verstraeten, G., Van Oost, K., Govers, G., and Poesen, J.: Modelling mean annual sediment yield using a distributed approach, Earth Surf. Proc. Land., 26, 1221–1236, https://doi.org/10.1002/esp.275, 2001.
}
