% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soilgrids.R
\name{get_soilgrids}
\alias{get_soilgrids}
\title{Download SoilGrids layers}
\usage{
get_soilgrids(aoi, layer = "all")
}
\arguments{
\item{aoi}{SpatVector. A polygon layer with area of interest.}

\item{layer}{character. A string indicating what layers should be downloaded. Either one of the following: \code{'all'}, \code{'sand'}, \code{'clay'}, \code{'silt'} or \code{'soc'}}
}
\value{
SpatRaster
}
\description{
Download SoilGrids layers
}
\examples{
library(terra)
library(purrr)

f <- system.file("extdata/extent.shp", package="rusleR")
v <- vect(f)

sand <- get_soilgrids(v, layer = "sand")

plot(sand)

}
