% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soilgrids.R
\name{get_soilgrids}
\alias{get_soilgrids}
\title{Download SoilGrids layers}
\usage{
get_soilgrids(
  aoi,
  layer = c("all", "sand", "silt", "clay", "soc", "phh2o", "bdod"),
  pred.quantile = "mean"
)
}
\arguments{
\item{aoi}{SpatVector. A polygon layer with area of interest.}

\item{layer}{character. A string indicating what layers should
be downloaded. Either one of the following: \code{'all'},
\code{'sand'}, \code{'clay'}, \code{'silt'}, \code{'soc'},
\code{bdod} or \code{'phh2o'}}

\item{pred.quantile}{character. A string indicating what predition
quantiles should be downloaded. Either one of the following: \code{'mean'}
(default), \code{'Q0.05'}, \code{'Q0.95'} or \code{'Q0.05'} (see Details).}
}
\value{
SpatRaster
}
\description{
A function to download \href{https://www.isric.org/explore/soilgrids}{SoilGrids}
dataset, created by \emph{Hengl et al.} (2017)
}
\examples{
library(terra)
library(purrr)

f <- system.file("extdata/extent.shp", package="rusleR")
v <- vect(f)

sand <- get_soilgrids(v, layer = "sand")

plot(sand)

}
\references{
Hengl, Tomislav, Jorge Mendes de Jesus, Gerard B. M. Heuvelink, Maria Ruiperez Gonzalez, Milan Kilibarda, Aleksandar Blagotić, Wei Shangguan, et al. “SoilGrids250m: Global Gridded Soil Information Based on Machine Learning.” PLOS ONE 12, no. 2 (February 16, 2017): e0169748. https://doi.org/10.1371/journal.pone.0169748.
}
